// Front/src/open-api/model/categoryDto.ts
/**
 * OpenApi specification
 * OpenApi documentation for Admin Module Demo
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CategoryDto {
    id?: number;
    name?: string;
    ruleCount?: number;
    editmode?:boolean;
}